Refactor account screen
